+++ TheBusanTrail.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32228.343
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33502.453
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheBusanTrail", "TheBusanTrail.csproj", "{AB13B15D-558F-4E36-B935-FF49FC59D27A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheBusanTrail", "TheBusanTrail\TheBusanTrail.csproj", "{644D83AF-D4B3-427B-A377-E135A9C028FC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AB13B15D-558F-4E36-B935-FF49FC59D27A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB13B15D-558F-4E36-B935-FF49FC59D27A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB13B15D-558F-4E36-B935-FF49FC59D27A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB13B15D-558F-4E36-B935-FF49FC59D27A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{644D83AF-D4B3-427B-A377-E135A9C028FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{644D83AF-D4B3-427B-A377-E135A9C028FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{644D83AF-D4B3-427B-A377-E135A9C028FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{644D83AF-D4B3-427B-A377-E135A9C028FC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A255EF46-525B-4515-9698-0D80D494D1A9}
+		SolutionGuid = {E49D6350-2A08-4D22-BDDF-CC88803BE24E}
 	EndGlobalSection
 EndGlobal
