+++ TheBusanTrail/Program.cs
@@ -1,0 +1,3 @@
+﻿
+using var game = new TheBusanTrail.Game1();
+game.Run();
